refactor: update in utils folder. Create new files for clear logic
